Search for direnv on PATH, expanding leading ~ in PATH components.
Bash and Python, for example, appear to expand entries on PATH like ~/bin into
something like /user/home/dir/bin, before searching for an unqualified
binary (e.g. "direnv") to execute. Rust does not seem to do this, which can be
confusing for users. Here we search PATH explicitly, expanding ~/bin and the
like as we go.