Emit logs to stderr, not stdout.
It's expected to have a line like the following in `.envrc`:

  eval "$(firstaide hook)"

If logs are going to stdout, then Bash will try to evaluate them. Equally, if
there's an error, Bash will complain about not being able to understand the
syntax, whereas we would prefer to see the log messages as they are.