On second thoughts, don't try to compile in the path to direnv.
For best compatibility with the end user's setup, firstaide *must* use the
direnv that the user is using in their shell hook, and must version check _that_
binary. Using a hard-coded path to a direnv in the Nix store actually makes
things worse because the user's shell hook may well not be pointing to the same
version of direnv.