For each build, update a `cache` symlink to point to it.
This acts as a fallback cache when a cache corresponding to the current file
sums does not exist. The fallback is to the most recently build cache, not
necessarily to some "closest" match because it's not clear how exactly we might
calculate closeness. Recentness is a reasonably proxy for now.